requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2024.8.30
charset-normalizer==2.1.1
idna==3.10
python-dotenv==0.21.0
python-telegram-bot==13.15
pytz==2024.2
requests==2.28.1
setuptools==75.6.0
six==1.17.0
tornado==6.1
tzlocal==5.2
urllib3==1.26.20
